+++ scripts/lichessTools-init.js
@@ -99,7 +99,7 @@ lichessTools.loadTool(LiChessTools.Tools.ForkBehavior);
 lichessTools.loadTool(LiChessTools.Tools.BlogAutosave);
 lichessTools.loadTool(LiChessTools.Tools.AutoAnalyseGame);
 lichessTools.loadTool(LiChessTools.Tools.ActiveIcon);
-lichessTools.loadTool(LiChessTools.Tools.ImagePasting);
+lichessTools.loadTool(LiChessTools.Tools.ChatForumBlogs);
 lichessTools.loadTool(LiChessTools.Tools.RemoveChatLinkWarning);
 lichessTools.loadTool(LiChessTools.Tools.MchatOptions);
 lichessTools.loadTool(LiChessTools.Tools.FullScreen);
@@ -131,6 +131,7 @@ lichessTools.loadTool(LiChessTools.Tools.ChallengeOptions);
 lichessTools.loadTool(LiChessTools.Tools.BetterBestArrow);
 lichessTools.loadTool(LiChessTools.Tools.DownloadBroadcastPgn);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerMoreGames);
+lichessTools.loadTool(LiChessTools.Tools.CustomBoardImage);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);