+++ tools/customEngineLevel/tool.js
@@ -90,14 +90,14 @@
             elem.replaceText(text=>{
               return text
                        .replace(/(\d+)(?:\/\d+)+/, '$1/' + customDepth);
-            });
+            }, true);
           } else {
             const reg = new RegExp('^('+pattern+')(?:\\/\\d+)?$');
             // lichess keeps a reference to the actual node
             elem.replaceText(text=>{
               return text
                        .replace(reg, '$1/' + customDepth);
-            });
+            }, true);
           }
         }
       }