+++ tools/playLayout/tool.js
@@ -8,6 +8,14 @@
         type: 'single',
         possibleValues: ['normal', 'noSide', 'smallSide', 'smallSideable'],
         defaultValue: 'normal'
+      },
+      {
+        name: 'playLayoutElements',
+        category: 'play',
+        type: 'multiple',
+        possibleValues: ['deadCrosstable'],
+        defaultValue: false,
+        advanced: true
       }
     ];
 
@@ -15,19 +23,23 @@
       'en-US': {
         'options.play': 'Play',
         'options.playLayout': 'Play layout',
+        'options.playLayoutElements': 'Play layout elements',
         'playLayout.normal': 'Normal',
         'playLayout.noSide': 'Hide left side',
         'playLayout.smallSide': 'Hide chat',
         'playLayout.smallSideable': 'Option to hide chat',
+        'playLayoutElements.deadCrosstable': 'Non interactive crosstable',
         'toggleLayoutTitle': 'LiChess Tools - click on the icon to hide/show chat'
       },
       'ro-RO': {
         'options.play': 'Joc',
         'options.playLayout': 'Aranjament vizual',
+        'options.playLayoutElements': 'Elemente vizuale',
         'playLayout.normal': 'Normal',
         'playLayout.noSide': 'Ascunde partea st\u00e2ng\u0103',
         'playLayout.smallSide': 'Ascunde chat',
         'playLayout.smallSideable': 'Op\u0163iune s\u0103 ascunzi chat',
+        'playLayoutElements.deadCrosstable': 'Tabel rezultate neinteractiv',
         'toggleLayoutTitle': 'LiChess Tools - apas\u0103 pictograma ca s\u0103 ascunzi/ar\u0103\u0163i chat'
       }
     }
@@ -37,7 +49,7 @@
       const $ = lt.$;
       if (!$(ev.target).is('.game__meta__infos')) return;
       ev.preventDefault();
-      switch (this.options.value) {
+      switch (this.options.playLayout) {
         case 'noSide':
           $('body').toggleClass('lichessTools-noSide');
           break;
@@ -60,7 +72,7 @@
         .removeClass('lichessTools-pointer')
         .removeAttr('title')
         .off('click', this.toggleLayout);
-      if (this.options.value != 'normal') {
+      if (this.options.playLayout != 'normal') {
         $('.playing main.round .game__meta__infos')
           .addClass('lichessTools-pointer')
           .attr('title', trans.noarg('toggleLayoutTitle'))
@@ -72,16 +84,22 @@
       const lt = this.lichessTools;
       const lichess = lt.lichess;
       const $ = lt.$;
-      const value = lt.currentOptions.getValue('playLayout');
-      this.options = { value: value };
-      this.logOption('Play layout', value);
+      const playLayout = lt.currentOptions.getValue('playLayout');
+      const playLayoutElements = lt.currentOptions.getValue('playLayoutElements');
+      this.options = { 
+        playLayout: playLayout,
+        deadCrosstable: lt.isOptionSet(playLayoutElements, 'deadCrosstable')
+      };
+      this.logOption('Play layout', playLayout);
+      this.logOption('Play layout elements', playLayoutElements);
       this._control = $('main.round .game__meta__infos')[0];
       if (!this._control) return;
       $('body')
-        .toggleClass('lichessTools-noSide', value == 'noSide')
-        .toggleClass('lichessTools-smallSide', value == 'smallSide');
+        .toggleClass('lichessTools-noSide', this.options.playLayout == 'noSide')
+        .toggleClass('lichessTools-smallSide', this.options.playLayout == 'smallSide')
+        .toggleClass('lichessTools-deadCrosstable', this.options.deadCrosstable);
       lt.global.clearInterval(this.interval);
-      if (value != 'normal') {
+      if (this.options.playLayout != 'normal') {
         this.interval = lt.global.setInterval(() => {
           const control = $('main.round .game__meta__infos')[0];
           if (this._control != control) {